Request existing groups and subjects related
When page is ready manager.js makes a request for the groups and subjets.
Returns a JSON with them.

When a request is made, manage_data.php pull the data from the database.
Return a JSON with 2 arrays, one for the groups with subjets assigned and one
for the groups with no subjets assigned.